Rename api module to rpc
